vacation <- jaewon/dev

+++ src/main/java/com/sparta/nexusteam/vacation/service/VacationServiceImpl.java
@@ -14,6 +14,7 @@ import com.sparta.nexusteam.vacation.entity.Vacation;
 import com.sparta.nexusteam.vacation.entity.VacationType;
 import com.sparta.nexusteam.vacation.repository.VacationRepository;
 import com.sparta.nexusteam.vacation.repository.VacationTypeRepository;
+import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -35,7 +36,7 @@ public class VacationServiceImpl implements VacationService {
     @Override
 //    @CacheEvict(value = "vacationTypes", key = "#companyId")
     public VacationTypeResponse createVacationType(PostVacationTypeRequest requestDto,
-            Long companyId) {
+                                                   Long companyId) {
         Company company = companyRepository.findById(companyId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 회사가 없습니다"));
         VacationType vacationType = new VacationType(requestDto.getName(), requestDto.getDays(),
@@ -47,9 +48,15 @@ public class VacationServiceImpl implements VacationService {
     @Override
 //    @CacheEvict(value = "vacations", key = "#employee.id")
     public VacationResponse createVacation(Long vacationTypeId, PostVacationRequest requestDto,
-            Employee employee) {
+                                           Employee employee) {
         VacationType vacationType = vacationTypeRepository.findById(vacationTypeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류가 없습니다."));
+        if(requestDto.getStartDate().isAfter(requestDto.getEndDate())) {
+            throw new IllegalArgumentException("시작일시는 종료일시 보다 이르면 안됩니다.");
+        }
+        if(vacationType.getDays()*24 < Duration.between(requestDto.getStartDate(), requestDto.getEndDate()).toHours()) {
+            throw new IllegalArgumentException(("시작일시와 종료일시의 차이는 휴가 종류 일수보다 크면 안 됩니다."));
+        }
         Vacation vacation = new Vacation(requestDto.getStartDate(), requestDto.getEndDate(),
                 vacationType, employee);
         vacation = vacationRepository.save(vacation);
@@ -99,7 +106,7 @@ public class VacationServiceImpl implements VacationService {
     @Transactional
 //    @CacheEvict(value = "vacation", key = "#vacationId")
     public VacationResponse updateVacationApprovalStatus(Long vacationId,
-            PatchVacationApprovalRequest requestDto) {
+                                                         PatchVacationApprovalRequest requestDto) {
         Vacation vacation = vacationRepository.findById(vacationId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가는 없습니다."));
         vacation.updateApprovalStatus(requestDto.getApprovalStatus());
@@ -112,13 +119,16 @@ public class VacationServiceImpl implements VacationService {
     public void deleteVacation(Long vacationId) {
         Vacation vacation = vacationRepository.findById(vacationId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가는 없습니다."));
+        if(vacation.getEndDate().isAfter(LocalDateTime.now())) {
+            throw new IllegalArgumentException("사용한 휴가는 삭제 할수 없습니다.");
+        }
         vacationRepository.delete(vacation);
     }
 
     @Override
     @Transactional
     public VacationTypeResponse updateVacationType(Long vacationTypeId,
-            PutVacationTypeRequest requestDto) {
+                                                   PutVacationTypeRequest requestDto) {
         VacationType vacationType = vacationTypeRepository.findById(vacationTypeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류는 없습니다."));
         vacationType.updateVacationType(requestDto);
@@ -132,4 +142,7 @@ public class VacationServiceImpl implements VacationService {
                 .orElseThrow(() -> new IllegalArgumentException("해당 ID의 휴가 종류는 없습니다."));
         vacationTypeRepository.delete(vacationType);
     }
+
+
+
 }
